+++ src/fswAlgorithms/effectorInterfaces/rwMotorTorque/rwMotorTorque.c
@@ -186,7 +186,7 @@ void Update_rwMotorTorque(rwMotorTorqueConfig *configData, uint64_t callTime, in
 
     /* store the output message */
     rwMotorTorques = ArrayMotorTorqueMsg_C_zeroMsgPayload();
-    vCopy(us, configData->rwConfigParams.numRW, rwMotorTorques.motorTorque);
+    vCopy(us, (size_t) configData->rwConfigParams.numRW, rwMotorTorques.motorTorque);
     ArrayMotorTorqueMsg_C_write(&rwMotorTorques, &configData->rwMotorTorqueOutMsg, moduleID, callTime);
 
     return;
